Conflicts:
	system/language/english/form_validation_lang.php
	user_guide/changelog.html
	user_guide/libraries/form_validation.html

+++ system/libraries/Cache/drivers/Cache_file.php
@@ -107,7 +107,14 @@ class CI_Cache_file extends CI_Driver {
 	 */
 	public function delete($id)
 	{
-		return unlink($this->_cache_path.$id);
+		if (file_exists($this->_cache_path.$id))
+		{
+			return unlink($this->_cache_path.$id);
+		}
+		else
+		{
+			return FALSE;
+		}
 	}
 
 	// ------------------------------------------------------------------------
@@ -157,17 +164,16 @@ class CI_Cache_file extends CI_Driver {
 		
 		if (is_array($data))
 		{
-			$data = $data['data'];
 			$mtime = filemtime($this->_cache_path.$id);
 
-			if ( ! isset($data['ttl']))
+			if ( ! isset($data['data']['ttl']))
 			{
 				return FALSE;
 			}
 
 			return array(
-				'expire' 	=> $mtime + $data['ttl'],
-				'mtime'		=> $mtime
+				'expire' => $mtime + $data['data']['ttl'],
+				'mtime'	 => $mtime
 			);
 		}
